encryption_server/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
pycryptodome==3.17
